% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA_polygon.R
\docType{data}
\name{CA_polygon}
\alias{CA_polygon}
\title{Sample polygon data of California counties.}
\format{
A \code{sf} object with 58 rows and 2 variables
\describe{
  \item{id}{FIPS County code of geography}
  \item{geometry}{the geometry column for counties(CRS: NAD83)}
}
}
\usage{
CA_polygon
}
\description{
A sf(simple feature) containing geometric boundaries of CA counties with their codes.
}
\keyword{datasets}
