% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gij.polygon.R
\name{Gij.polygon}
\alias{Gij.polygon}
\title{Calculate spatial autocorrelation with OD data and corresponding polygons.}
\usage{
Gij.polygon(df, shape, queen = TRUE, snap = 1, method = "t", n = 1000)
}
\arguments{
\item{df}{A data.frame that contains your Origin-Destination data. The df must consist of "oid"(origin id), "did"(destination id), "n"(flow value).}

\item{shape}{A shapefile(in a polygon type) that matches to your OD data.frame. The shape must have "id" column to match your df.}

\item{queen}{A T/F input that is used to calculate spdep's spatial contingency(Please view documents of spdep package for more information).}

\item{snap}{A parameter that is also used to calculate spdep's spatial contingency(Please view documents of spdep package for more information).}

\item{method}{A character value among "o"(origin based), "d"(destination based), and "t"(both way) which determines the way to generate Spatial Weights. The default value is "t".}

\item{n}{A integer value to define how many times you want to execute bootstrapping.}
}
\value{
The result is in the form of a list which includes dataframe and lines.
The first result consists of Gij statistics and p-value in columns merged to your input df.
The latter has the same attributes of the former, whereas its form is lines.
}
\description{
Calculate spatial autocorrelation with OD data and corresponding polygons.
}
\examples{
\dontrun{
result <- Gij.polygon(df = Grid_OD, shape = Grid, queen = TRUE, snap = 1, method = 't', n = 1)
result[[1]] # data.frame
result[[2]] # lines
}
}
